refactor: divide InputOutput class into Input and Output class
